move logic to analytics store
